Now the application won't put the data entered by the user in the checkin sheet, rather it would put the matched data that was found in the signup sheet.
Suppose in Sign Up Sheet we had

Sahib Grover 8582813471

If we entered just 8582813471, it would find me, but in the Check in Sheet it would put in : Unspecified Unspecified 8582813471.

Nowe if someone came and checked in using: Sahib Grover, it would find me in the signup sheet, but wont match me in the check in sheet because there is no mention of Sign Up sheet.

To deal with this bug, even if you just enter your phone number, the checkin sheet will now have the first name, lastName and the phone number. :)